[JSC] Avoid saving top-level scope for non-module code https://bugs.webkit.org/show_bug.cgi?id=252674 rdar://105732218
Reviewed by NOBODY (OOPS!).

Avoid allocating a local variable and moving the scope to save the top-level
scope. Re-loading the context is pretty cheap and this reduces the amount of
bytecode overhead for small functions. Notice that we still allocate a variable
for module code, which is necessary since we allocate a lexical environment,
which needs the scope for variable resolution.

* JSTests/stress/hoisted-eval-different-scope.js: Added.
(i.with.g):
(i.g.f):
(i.g):
* Source/JavaScriptCore/bytecompiler/BytecodeGenerator.cpp:
(JSC::BytecodeGenerator::BytecodeGenerator):
(JSC::BytecodeGenerator::emitResolveScopeForHoistingFuncDeclInEval):
(JSC::BytecodeGenerator::allocateAndEmitScope):
(JSC::BytecodeGenerator::restoreScopeRegister):