[WGSL] wgslc should compile all entry points by default https://bugs.webkit.org/show_bug.cgi?id=290517 rdar://147999783
Reviewed by NOBODY (OOPS!).

Previously, when passing `_` as the entry point name to wgslc it ignored the error
that entry point does not exist, but it also did not compile any entry points.
Change the behavior so all entry points are compiled instead, since that information
is already available in the call graph. Also makes the `_` argument optional, so
invoking it with `wgslc <file>` compiles all entry points by default.

* Source/WebGPU/WGSL/wgslc.cpp:
(printUsageStatement):
(CommandLine::parseArguments):
(runWGSL):