[WGSL] Can't pack/unpack nested arrays https://bugs.webkit.org/show_bug.cgi?id=290536 rdar://147752033
Reviewed by NOBODY (OOPS!).

The signature of the packing helpers for arrays only worked for vec3 and structs,
but not for nested arrays. To address the recursion we introduce two helper types
`__UnpackedType<T>` and `__PackedType<T>` to the mapping from T to their packed/
unpacked counterparts.

* Source/WebGPU/WGSL/Metal/MetalFunctionWriter.cpp:
(WGSL::Metal::FunctionDefinitionWriter::emitNecessaryHelpers):
(WGSL::Metal::FunctionDefinitionWriter::visit):
* Source/WebGPU/WGSL/tests/valid/packing-nested-array.wgsl:
* Source/WebGPU/WGSL/tests/valid/packing.wgsl: